feat: #6
Implemented the 'loadFloor' function in the utility file to generate a grid of floor planes within the Three.js scene. Each floor plane has custom dimensions defined by WIDTH and LENGTH constants and is repeated in a grid pattern specified by NUM_X and NUM_Z constants. The floor planes are created with a standard mesh material with custom properties like displacementScale and roughness. Additionally, each floor plane is configured to receive shadows but not cast them. The function enhances the scene by adding visual depth and realism through the representation of the floor surface.

Co-Authored-By: Faru Nuri Sönmez <info@farunurisonmez.com>