#10
Created the ModesSidebarComponent to facilitate navigation within the application. This component includes a toggle button that opens and closes the sidebar, providing access to different sections of the app. The sidebar menu items are dynamically rendered based on the user's permissions and the current application state. Additionally, implemented transition animations for smooth opening and closing of the sidebar, enhancing the overall user experience."

Co-Authored-By: Faru Nuri Sönmez <info@farunurisonmez.com>